refactor: drop rag_service shim
