+++ rag_demo/app/app/scripts/rag_optuna_tune.py
@@ -6,7 +6,7 @@ from typing import Dict, Any, List, Tuple
 
 import optuna  # pip install optuna
 
-from app.app.services.rag_service import RagService
+from app.app.services.rag import RagService
 from app.app.metrics.quality import (
     keys_from_docs, hit_at_k, recall_at_k, dup_rate, p_percentile
 )
