--- rag_demo/app/app/services/rag_service.py
@@ -1,6 +1,0 @@
-"""Backward compatible import for RagService."""
-
-from .rag.service import RagService
-
-__all__ = ["RagService"]
-
